dnallm/mcp/requirements.txt
# MCP Server Dependencies
mcp>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.10.6
pyyaml>=6.0.1
asyncio>=3.4.3
aiofiles>=23.2.1
python-multipart>=0.0.6
httpx>=0.25.0

# Existing DNALLM dependencies (already in pyproject.toml)
# torch>=2.4.0,<2.7
# transformers>=4.49.0
# accelerate>=1.4.0
# datasets<=3.2.0
# huggingface-hub>=0.29.0
# scikit-learn>=1.4.0
# numpy<2
# pandas>=2.2.3
